+++ fake-cucumber/javascript/src/IncrementClock.ts
@@ -1,9 +1,0 @@
-import IClock from './IClock'
-
-export default class IncrementClock implements IClock {
-  private time = 0
-
-  public clockNow(): number {
-    return this.time++
-  }
-}