+++ compatibility-kit/CHANGELOG.md
@@ -17,6 +17,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [10.0.0] - 2022-06-09
+
+### Changed
+
+* Update the ndjson files in the kit using `fake-cucumber` 16.0.0 to bring new
+  keyword types to steps and pickles
+  (refs. [#1966](https://github.com/cucumber/common/pull/1966)
+  [#1741](https://github.com/cucumber/common/pull/1741))
+
 ## [9.2.1] - 2022-05-19
 
 ### Fixed
@@ -136,7 +145,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/common/compare/compatibility-kit/v9.2.1...main
+[Unreleased]: https://github.com/cucumber/common/compare/compatibility-kit/v10.0.0...main
+[10.0.0]:      https://github.com/cucumber/common/compare/compatibility-kit/v9.2.1...v10.0.0
 [9.2.1]:      https://github.com/cucumber/common/compare/compatibility-kit/v9.2.0...v9.2.1
 [9.2.0]:      https://github.com/cucumber/common/compare/compatibility-kit/v9.1.2...v9.2.0
 [9.1.2]:      https://github.com/cucumber/common/compare/compatibility-kit/v9.1.1...v9.1.2